requirements.txt
mitmproxy==6.0.2
Pillow==8.2.0
qrcode==6.1
